added print css, added document status, configure usage of Google Analytics
print css:
css file for media type print with no sidebar and header

document status:
currently configured are status inprogress and finished, they are shown
in the document list as well as in the menu.

configure Google Analytics
if `gaCode` is set to a Google Analytics identifier in `settings.json`
the necessary script is added to all generated pages. In the default it
is blank, so no script will be rendered.